requirement.txt
django
djangorestframework
celery
redis
psycopg2-binary
ragas
langchain
langchain-community 
pytest
pytest-django 
coverage
dotenv